Restore using await_for_async in storageService.py
This reverts d565dbf.  It can take a long time for the Archivematica
storage service to successfully store an AIP in the Wellcome storage,
because:

1. We unpack and repack the bag with the new External-Identifier
2. The Wellcome storage service has to scale up to store a new bag,
   which means it can take a long time to store a bag.

Long term it would be nice to not diverge from the Archivematica
codebase here, but unless we can speed up the storage service this is
unlikely.