Compute checksums *before* writing to the DB
In the "Assign file UUIDs and checksums" step, Archivematica would
previously try to compute the checksum and size of every file while
holding open an atomic database transaction.  Our biggest accession is
100 files totalling ~180GB, with several files that are ~17GB each.
We saw some issues at this step, for example

    OperationalError: (1213, ‘Deadlock found when trying to get lock;
    try restarting transaction’)

in the MCP client logs.

This patch modifies MCP client to compute all the checksums *before* it
writes anything to the database, so the atomic transaction is only held
open for a short window.